sketch for Perry Initiative
This is a sketch I developed for the Perry Initiative. It was used for a contest where young women attempted to repair a broken fake bone. It was then subjected to a load until failure. This sketch measured the applied load and printed the max force when the bone broke.